cypress + percy for each storybook component
For each storybook component, creates an initial cypress test that loads the component in its storybook iframe and takes a percy snapshot.

This replaces percy-storybook, which did not wait for the Proxima Nova webfont to load. We'll see if cypress out of the box does better, but even if not we'll have more tools to fix it -- percy-storybook renders screenshots on percy servers and is essentially out of our control.

We can also use these for our unit testing, testing additional functionality where needed.

Initially we only test & screenshot the `Default` story for each component. This will save Percy screenshots, and we can change this as we deem  appropriate.