fix: window.innerWidth in Chrome devtools
Fixes a fairly obscure issue where the viewport width (used to determine which header to show) was incorrect when the responsive window size was set to 1024px or smaller in Chrome devtools.

I was unable to reproduce the issue elsewhere and don't understand the full implications of this change, so I don't know if it's worth making. But as far as I can tell everything looks good.